修复了中文路径下创建快捷启动失败的问题

+++ build.bat
@@ -2,7 +2,7 @@
 
 :: 清空 dist 文件夹（如果存在）
 if exist "dist" rmdir /S /Q "dist"
-rmdir /S /Q "build"
+if exist "build" rmdir /S /Q "build"
 del /Q "*.spec"
 
 :: 使用 PyInstaller 创建正常版本（无窗口）
@@ -10,7 +10,7 @@ pyinstaller --name="微信多开管理器" --windowed --icon=external_res\SunnyM
 --add-data="external_res;external_res" --manifest=app.manifest --exclude-module=PyQt5 --exclude-module=numpy ^
 --noconfirm Main.py > build_log_正式版.txt 2>&1
 
-:: 重命名生成的 EXE 文件
+:: 检查打包是否成功
 if exist "dist\微信多开管理器\微信多开管理器.exe" (
     echo 正式版打包成功
 ) else (
@@ -23,6 +23,7 @@ pyinstaller --name="微信多开管理器_调试版" --icon=external_res\SunnyMu
 --add-data="external_res;external_res" --manifest=app.manifest --exclude-module=PyQt5 --exclude-module=numpy ^
 --noconfirm Main.py > build_log_调试版.txt 2>&1
 
+:: 检查调试版打包是否成功
 if exist "dist\微信多开管理器_调试版\微信多开管理器_调试版.exe" (
     echo 调试版打包成功！
 
@@ -40,6 +41,7 @@ if exist "dist\微信多开管理器_调试版\微信多开管理器_调试版.e
 rmdir /S /Q "build"
 del /Q "*.spec"
 
+:: 复制快捷方式创建脚本到打包文件夹
 copy "点我创建快捷方式.bat" "dist\微信多开管理器\"
 
 echo 打包完成！正式版和调试版都已生成，共用同一个 _internal 文件夹。
